pyproject.toml
[tool.poetry]
name = "ic"
version = "0.0.1"
description = "Site of GameCenter"
authors = ["kill-your-soul <Platonovdi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.2"}
pillow = "^10.0.0"
psycopg2 = "^2.9.7"
django-environ = "^0.11.2"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^4.1.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
